Consolidate lox-py and lox-space crates
