+++ .github/workflows/python.yml
@@ -33,7 +33,7 @@ jobs:
         uses: PyO3/maturin-action@v1
         with:
           target: ${{ matrix.target }}
-          working-directory: ./crates/lox-py
+          working-directory: ./crates/lox-space
           args: --release --out dist --find-interpreter
           sccache: 'true'
           manylinux: auto
@@ -41,7 +41,7 @@ jobs:
         uses: actions/upload-artifact@v3
         with:
           name: wheels
-          path: ./crates/lox-py/dist
+          path: ./crates/lox-space/dist
 
   windows:
     runs-on: windows-latest
@@ -58,14 +58,14 @@ jobs:
         uses: PyO3/maturin-action@v1
         with:
           target: ${{ matrix.target }}
-          working-directory: ./crates/lox-py
+          working-directory: ./crates/lox-space
           args: --release --out dist --find-interpreter
           sccache: 'true'
       - name: Upload wheels
         uses: actions/upload-artifact@v3
         with:
           name: wheels
-          path: ./crates/lox-py/dist
+          path: ./crates/lox-space/dist
 
   macos:
     runs-on: macos-latest
@@ -81,14 +81,14 @@ jobs:
         uses: PyO3/maturin-action@v1
         with:
           target: ${{ matrix.target }}
-          working-directory: ./crates/lox-py
+          working-directory: ./crates/lox-space
           args: --release --out dist --find-interpreter
           sccache: 'true'
       - name: Upload wheels
         uses: actions/upload-artifact@v3
         with:
           name: wheels
-          path: ./crates/lox-py/dist
+          path: ./crates/lox-space/dist
 
   sdist:
     runs-on: ubuntu-latest
@@ -97,14 +97,14 @@ jobs:
       - name: Build sdist
         uses: PyO3/maturin-action@v1
         with:
-          working-directory: ./crates/lox-py
+          working-directory: ./crates/lox-space
           command: sdist
           args: --out dist
       - name: Upload sdist
         uses: actions/upload-artifact@v3
         with:
           name: wheels
-          path: ./crates/lox-py/dist
+          path: ./crates/lox-space/dist
 
   release:
     name: Release
@@ -120,6 +120,6 @@ jobs:
         env:
           MATURIN_PYPI_TOKEN: ${{ secrets.PYPI_API_TOKEN }}
         with:
-          working-directory: ./crates/lox-py
+          working-directory: ./crates/lox-space
           command: upload
           args: --non-interactive --skip-existing *
