+++ .github/workflows/rust.yml
@@ -32,7 +32,7 @@ jobs:
         uses: PyO3/maturin-action@v1
         with:
           command: develop
-          working-directory: ./crates/lox-py
+          working-directory: ./crates/lox-space
           args: --extras dev
       - name: Run test
         run: |
