+++ Cargo.toml
@@ -21,16 +21,17 @@ lox-core = { path = "crates/lox-core" }
 lox-earth = { path = "crates/lox-earth" }
 lox-ephem = { path = "crates/lox-ephem" }
 lox-io = { path = "crates/lox-io" }
-lox-py = { path = "crates/lox-py" }
 lox-space = { path = "crates/lox-space" }
 lox-time = { path = "crates/lox-time" }
 
+divan = "0.1.14"
 dyn-clone = "1.0.17"
 fast_polynomial = "0.1.0"
 float_eq = "1.0.1"
 glam = "0.25.0"
 nom = "7.1.3"
 proptest = "1.4.0"
+pyo3 = "0.20.3"
 rstest = "0.18.2"
 thiserror = "1.0"
