--- crates/lox-py/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "lox_py"
-version = "0.1.0"
-
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-rust-version.workspace = true
-
-[lib]
-name = "lox_space"
-crate-type = ["cdylib"]
-
-[dependencies]
-lox-bodies.workspace = true
-lox-coords.workspace = true
-lox-core.workspace = true
-lox-time.workspace = true
-
-pyo3 = "0.20.2"
-float_eq.workspace = true
-thiserror.workspace = true
-
-[dev-dependencies]
-pyo3 = { version = "0.20.2", features = ["auto-initialize"] }
-rstest.workspace = true
