+++ crates/lox-space/Cargo.toml
@@ -1,13 +1,35 @@
 [package]
 name = "lox-space"
-version = "0.0.0"
+version = "0.1.0"
 
 authors.workspace = true
 edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
 
+[lib]
+name = "lox_space"
+crate-type = ["lib", "cdylib"]
+
 [dependencies]
 lox-bodies.workspace = true
 lox-coords.workspace = true
+lox-core.workspace = true
 lox-time.workspace = true
+
+float_eq.workspace = true
+pyo3.workspace = true
+thiserror.workspace = true
+
+[dev-dependencies]
+pyo3 = { version = "0.20.2", features = ["auto-initialize"] }
+divan.workspace = true
+rstest.workspace = true
+
+[[bench]]
+name = "iau_frames"
+harness = false
+
+[[bench]]
+name = "time_delta"
+harness = false
