+++ crates/lox-space/examples/iss.rs
@@ -6,9 +6,10 @@
  * file, you can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
+use lox_bodies::Earth;
+use lox_coords::DVec3;
 use lox_coords::frames::Icrf;
 use lox_coords::two_body::{Cartesian, Keplerian};
-use lox_space::prelude::*;
 use lox_time::continuous::{TDB, Time};
 use lox_time::continuous::julian_dates::JulianDate;
 use lox_time::dates::Date;
