+++ crates/lox-space/src/coords.rs
