--- crates/lox-space/src/prelude.rs
@@ -1,11 +1,0 @@
-/*
- * Copyright (c) 2023. Helge Eichhorn and the LOX contributors
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, you can obtain one at https://mozilla.org/MPL/2.0/.
- */
-
-pub use lox_bodies::*;
-pub use lox_coords::*;
-pub use lox_time::*;
